cleaning the workflows
